+++ addon/components/object-list-view.js
@@ -46,13 +46,14 @@ export default FlexberryBaseComponent.extend(FlexberryLookupCompatibleComponentM
       this.sendAction(action, column);
     },
 
-    deleteRow: function(key, record) {
-      if (this.get('readonly')) {
+    // TODO: rename recordWithKey. rename record in the template, where it is actually recordWithKey.
+    deleteRow: function(recordWithKey) {
+      if (this.get('readonly') || !recordWithKey.config.canBeDeleted) {
         return;
       }
 
       if (confirm('Do you really want to delete this record?')) {
-        this._deleteRecord(record, this.get('immediateDelete'));
+        this._deleteRecord(recordWithKey.data, this.get('immediateDelete'));
       }
     },
 
@@ -509,6 +510,52 @@ export default FlexberryBaseComponent.extend(FlexberryLookupCompatibleComponentM
   filterByAnyMatch: 'filterByAnyMatch',
 
   /**
+   * Hook for configurate rows.
+   *
+    Example:
+    ```handlebars
+    <!-- app/templates/employees.hbs -->
+    {{flexberry-objectlistview
+      ...
+      configurateRow=(action 'configurateRow')
+      ...
+    }}
+    ```
+
+    ```js
+    // app/controllers/employees.js
+    import ListFormController from './list-form';
+
+    export default ListFormController.extend({
+      actions: {
+        configurateRow: function(rowConfig, record) {
+          rowConfig.canBeDeleted = false;
+        }
+      }
+    });
+    ```
+   * @method configurateRow
+   * @param {Object} config Settings for row.
+                            See {{#crossLink "ObjectListView/defaultRowConfig:property"}}{{/crossLink}}
+                            property for details.
+   * @param {DS.Model} record The record in row.
+   */
+  configurateRow: undefined,
+
+  /**
+   * Default settings for rows.
+   *
+   * @property defaultRowConfig
+   * @type Object
+   * @param {Boolean} [canBeDeleted=true] The row can be deleted.
+   * @param {Boolean} [canBeSelected=true] The row can be selected via checkbox.
+   */
+  defaultRowConfig: {
+    canBeDeleted: true,
+    canBeSelected: true
+  },
+
+  /**
    * Flag: indicates whether DELETE request should be immediately sended to server (on each deleted record) or not.
    *
    * @property immediateDelete
@@ -751,6 +798,16 @@ export default FlexberryBaseComponent.extend(FlexberryLookupCompatibleComponentM
 
     modelWithKey.set('key', key);
     modelWithKey.set('data', record);
+
+    let rowConfig = Ember.copy(this.get('defaultRowConfig'));
+    modelWithKey.set('config', rowConfig);
+
+    let configurateRow = this.get('configurateRow');
+    if (configurateRow) {
+      Ember.assert('configurateRow must be a function', typeof configurateRow === 'function');
+      configurateRow(rowConfig, record);
+    }
+
     this.get('contentWithKeys').pushObject(modelWithKey);
 
     return key;
@@ -809,7 +866,60 @@ export default FlexberryBaseComponent.extend(FlexberryLookupCompatibleComponentM
     }
   },
 
+  /**
+   * Hook that executes before deleting the record.
+   *
+   *  Example:
+   *  ```handlebars
+   *  <!-- app/templates/employees.hbs -->
+   *  {{flexberry-objectlistview
+   *    ...
+   *    beforeDeleteRecord=(action 'beforeDeleteRecord')
+   *    ...
+   *  }}
+   *  ```
+   *
+   *  ```js
+   *  // app/controllers/employees.js
+   *  import ListFormController from './list-form';
+   *
+   *  export default ListFormController.extend({
+   *    actions: {
+   *      beforeDeleteRecord: function(record, data) {
+   *        if (record.get('myProperty')) {
+   *          data.cancel = true;
+   *        }
+   *      }
+   *    }
+   *  });
+   *  ```
+   *
+   * @method beforeDeleteRecord
+   *
+   * @param {DS.Model} record Deleting record.
+   * @param {Object} data Metadata.
+   * @param {Boolean} [data.cancel=false] Flag for canceling deletion.
+   * @param {Boolean} [data.immediately] See {{#crossLink "ObjectListView/immediateDelete:property"}}{{/crossLink}}
+   *                                     property for details.
+   */
+  beforeDeleteRecord: null,
+
   _deleteRecord: function(record, immediately) {
+    let beforeDeleteRecord = this.get('beforeDeleteRecord');
+    if (beforeDeleteRecord) {
+      Ember.assert('beforeDeleteRecord must be a function', typeof beforeDeleteRecord === 'function');
+
+      let data = {
+        immediately: immediately,
+        cancel: false
+      };
+      beforeDeleteRecord(record, data);
+
+      if (data.cancel) {
+        return;
+      }
+    }
+
     var key = this._getModelKey(record);
     this._removeModelWithKey(key);