+++ tests/dummy/app/router.js
@@ -12,6 +12,9 @@ Router.map(function() {
   // /employees
   this.route('employees');
 
+  // /employee-edit - route for test custom edit form route in f-olv
+  this.route('employee-edit', { path: 'employee-edit/:id' });
+
   // /employees/2 - render into outlet in application template
   this.route('employee', { path: 'employees/:id' });