+++ tests/dummy/app/routes/employee-edit.js
@@ -1,0 +1,10 @@
+import EditFormRoute from './edit-form';
+import AuthenticatedRouteMixin from 'ember-simple-auth/mixins/authenticated-route-mixin';
+
+export default EditFormRoute.extend(AuthenticatedRouteMixin, {
+  modelProjection: 'EmployeeE',
+  modelName: 'employee',
+
+  controllerName: 'employee',
+  templateName: 'employee'
+});